QintelQwatch: Renaming qintel_helper file
