+++ qintelqwatch_helper.py
